pyproject.toml
[tool.poetry]
name = "crs-backend"
version = "1.0.0"
description = "hotel booking manager"
authors = ["osw4l <ioswxd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
factory-boy = "^3.2.1"
drf-yasg = "^1.21.4"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.5"
psycopg2 = "^2.9.5"
elasticsearch = "7.12.0"
elasticsearch-dsl = "7.2.0"
django-elasticsearch-dsl = "7.2.0"
Django = "3.2"
django-extensions = "^3.2.1"

[tool.poetry.dev-dependencies]
coverage = "^6.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

